+++ uRADMonitorX/Commons/Pressure.cs
@@ -4,16 +4,16 @@ namespace uRADMonitorX.Commons {
 
     public static class Pressure {
 
-        public static decimal PascalToKiloPascal(decimal pascal) {
-            decimal kPascal = (pascal / 1000);
+        public static double PascalToKiloPascal(double pascal) {
+            double kPascal = (pascal / 1000);
             return kPascal;
         }
 
-        public static decimal PascalToAtm(decimal pascal) {
+        public static double PascalToAtm(double pascal) {
             throw new NotImplementedException();
         }
 
-        public static decimal PascalToBar(decimal pascal) {
+        public static double PascalToBar(double pascal) {
             throw new NotImplementedException();
         }
     }
