+++ uRADMonitorX/Commons/Temperature.cs
@@ -5,22 +5,22 @@ namespace uRADMonitorX.Commons {
 
     public static class Temperature {
 
-        public static decimal CelsiusToFahrenheit(decimal celsius) {
-            decimal fahrenheit = (celsius * 9 / 5) + 32;
+        public static double CelsiusToFahrenheit(double celsius) {
+            double fahrenheit = (celsius * 9 / 5) + 32;
             return fahrenheit;
         }
 
-        public static decimal FahrenheitToCelsius(decimal fahrenheit) {
-            decimal celsius = (fahrenheit - 32) * 5 / 9;
+        public static double FahrenheitToCelsius(double fahrenheit) {
+            double celsius = (fahrenheit - 32) * 5 / 9;
             return celsius;
         }
 
-        public static decimal CelsiusToKelvin(decimal celsius) {
-            return celsius + Decimal.Parse("273.15", System.Globalization.NumberStyles.AllowDecimalPoint, new CultureInfo("en-US").NumberFormat);
+        public static double CelsiusToKelvin(double celsius) {
+            return celsius + Double.Parse("273.15", System.Globalization.NumberStyles.AllowDecimalPoint, new CultureInfo("en-US").NumberFormat);
         }
 
-        public static decimal KelvinToCelsius(decimal kelvin) {
-            return kelvin - Decimal.Parse("273.15", System.Globalization.NumberStyles.AllowDecimalPoint, new CultureInfo("en-US").NumberFormat);
+        public static double KelvinToCelsius(double kelvin) {
+            return kelvin - Double.Parse("273.15", System.Globalization.NumberStyles.AllowDecimalPoint, new CultureInfo("en-US").NumberFormat);
         }
     }
 }
