+++ uRADMonitorX/Core/Device/DeviceData.cs
@@ -6,8 +6,8 @@ namespace uRADMonitorX.Core.Device {
 
         public DeviceInformation DeviceInformation { get; set; }
         public int Radiation { get; set; }
-        public decimal RadiationAverage { get; set; }
-        public decimal Temperature { get; set; }
+        public double RadiationAverage { get; set; }
+        public double Temperature { get; set; }
         public int? Pressure { get; set; }
         public int Voltage { get; set; }
         public int VoltagePercent { get; set; }
