+++ uRADMonitorX/Core/Device/DeviceReadings.cs
@@ -3,8 +3,8 @@
     public class DeviceReadings {
 
         public int Radiation { get; set; }
-        public decimal RadiationAverage { get; set; }
-        public decimal Temperature { get; set; }
+        public double RadiationAverage { get; set; }
+        public double Temperature { get; set; }
         public int? Pressure { get; set; }
         public int Voltage { get; set; }
         public int VoltagePercent { get; set; }
